requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic
python-dotenv
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2